Add a screenshot button
Added a button to take a screenshot of the current scene and copy it to
the clipboard.

This is somewhat jank right now, because I'm not properly checking
whether any images are still loading, so I just added a delay that
seems to be long enough for images to reliably appear in the screenshot.

Also, images must support CORS, or a screenshot cannot be taken. Status
icons from XIVAPI are automatically rewritten to point to the beta API,
which does provide CORS headers.