Improve arena presets
Renamed the "Savage" category to "Raids" since many of these presets
can be used in the normal versions of each instance as well.

Adjusted several presets for ultimate raids.

Added a preset for a unique arena in Dawntrail's second extreme trial.

Added a system where presets tied to specific instances are blurred and
may have names adjusted or redacted to avoid spoilers. Clicking a preset
or checking the "show all presets" box will reveal the preset's name and
preview.