Support hiding objects
All objects now have an opacity slider and a show/hide button in their
properties, as well as a show/hide button in the scene panel. Objects
that are hidden are no longer interactable except for their control
points when selected. Possible use cases include attaching a tether to
a point on the scene without any other object there and creating an
array of objects and carrying them across multiple steps, where
different subsets are shown on each step.

The H key can also be used to toggle the visibility of selected objects.

This also fixes some unrelated hotkeys including those for tethers.