Restructure theme code
Moved the scene theme code up to the root folder and consolidated it
with the Fluent UI theme code. Split it into separate theme objects for
the scene and for other UI elements, so the scene theme can later be set
by the current scene file while the other values can be set according to
the dark/light theme setting.

Pulled some things out to utility functions to make the theme
definitions easier to read.