Add text style options
Added options to switch between outline, drop shadow, and no additional
contrast for text objects. When set to use a shadow, the outline color
controls the shadow color instead. When the outline is disabled, the
outline color control is also disabled.

Also adjusted the outline size for text so it doesn't get too thin to be
usable for small fonts and doesn't get quite as thick for large fonts.