Improve tower apperance and controls
Updated the meteor/tower zone so that there is some separation between
the pillars, and the pillars for a 3-stack are arranged in a triangle.

Renamed "soak count" to "player count" and changed it from a numeric
input to a segmented control to make it easier to select the count.