Let dragging an object update the entire scene.
This should have no visible effect for most cases, but will cause any
tethers attached to the dragged object to update in real time rather than
only at the end.

Instead of using a temporary "drag position" to skip history updates,
explicitly signal to the UndoRedoReducer that the history should not be
updated on any drag event except onDragStart. This technically means
there's a no-op entry in the history until onDragMove is called, but I
couldn't find a way to trigger such a scenario (click+hold doesn't trigger
onDragStart).