Add support for "transient" updates
This takes the solution from #62 and implements it in a more general
manner so it can be applied to all kinds of state updates.

States can now be marked as "transient", which allows them to update the
state without updating the undo history. The state prior to a transient
update is tracked as the "canonical" state, so transient updates don't
affect things like the unsaved changes indicator. Applying a "commit"
action promotes the latest transient state to canonical and adds it to
the undo history, while applying a "rollback" action reverts to the
canonical state.

This allows for manipulating one object to also affect other objects in
real time, but for the entire action to still be counted as a single
change when using undo. For now, this new system is only used when
dragging objects, but it could be applied elsewhere later.