Highlight ('spotlight') the object in the scene when hovering over its entry in the scene list
The regular selection styling (including 'resizer') still takes
precedent if the object to spotlight is the only selected object, but
this will override the regular selection highlight if multiple objects
are selected (e.g. to help with finding one thing to exclude after
Ctrl+A).

The setup is parallel to the Selection setup, to avoid having to
refactor the selection setup itself (and having to come up with names
that read better than yielding `selection.selection`). It also uses a
Set to hold the object id to highlight, even though there will be at
most one entry. I started with making the `onMouseEnter` and
`onMouseLeave` events add/remove the item instead of setting the value
directly, but quickly found that some of the `onMouseLeave` events are
omitted when jiggling the mouse cursor. I kept the Set afterwards so I
didn't have to create a new structure just to not have to use a
placeholder value for "nothing needs the spotlight".

Along the way, make Tethers use the same setup as everything else,
instead of manually checking the selection set.

I picked the spotlight highlight colors in a way that they looked
distinguishable from both 'unselected' and 'selected' states for most
objects, but there are probably better color combinations available.