Simplify undo logic
The action types for undo, redo, etc. are now exported directly, and the
dispatch function includes those in the accepted action types. This
eliminates the need for extra hooks to access functions that call the
dispatch function with the hidden action types.

Now, instead of needing all of these for different tasks...

    const [undo, redo] = useSceneUndoRedo();
    const commit = useCommit();
    const rollback = useRollback();

...we can simply use "dispatch" from "useScene" for everything:

    const { dispatch } = useScene();
    dispatch({ type: 'undo' });
    dispatch({ type: 'redo' });
    dispatch({ type: 'commit' });
    dispatch({ type: 'rollback' });