Adjust spotlight color, fix faint highlights on some objects
Adjusted the spotlight color to yellow, since red felt a bit too much
like an error state, and there are more types of objects that default to
being red than yellow.

Also adjusted the highlight rendering for waymarks, look aways, stacks,
line stacks, proximity AOEs, and rotating AOEs so they are not fainter
than other objects.

Tethers and drawing objects still have faint outlines when they are made
transparent, but fixing that will require a larger change, so that is
being tracked in #66 and will be fixed separately.