Support dragging multiple objects with mouse input
Dragging an object that is in the selection will now move all objects in
the selection. Due to how Konva's dragging works, the objects that are
not the original target of the drag will lag behind by one frame, but
that usually isn't that noticeable, so it will be fixed later.

Fixes #18