Remove the option to select the parent object via a dropdown, and replace it with just displaying the parent object.
The dropdown could possibly be useful for e.g. quickly switching which player something is attached to, but 2 ways to select items (scene & scene list) should be sufficient for now and I couldn't get the styling right (it's sightly too wide, and triggers a horizontal scroll bar when there's also a vertical scroll bar). This element has also moved to below the X/Y row, so the "link" button doesn't change position after clicking it any more either. Double-click and then selecting the new parent from the scene list should be just as easy as the dropdown was.

Hovering over this element will trigger the spotlight on the parent object.