Let subsequent attachment of objects with a default attachment positon 'chain'.
This assumes that all still-pinned attachments are in their default positions.
In an ideal scenario there could be a "fully managed" mode where unpinning is not possible and deleting attachments would make the rest rearrange themselves, but I don't think that's worth the complexity in trying to explain in the UI (nor the complexity to implement) right now.
(maybe a dedicated "status effects" component that could also handle multi-row displays?)

Also reword some functions and arguments for clarity.