Strawman actually-make-the-objects-rotate implementation. the objects themselves work, but not the control points.
Not 100% yet sure if this is the right approach, or if the objects should only ever have absolute rotation values stored. This approach means relative rotation & position are easy to display and explain, and there's no need to complicate the 'update' action to update anything that might depend on the updated objects (and those that depend on those, etc) However this does mean that each object with special rendering logic needs to know to not read the x/y/rotation values directly (plus the control points/etc need to be more complicated).